requirement.txt
appdirs==1.4.4
asgiref==3.4.1
Brlapi==0.6.6
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.3.7
cx-Oracle==8.3.0
defer==1.0.6
defusedxml==0.5.0
distlib==0.3.4
Django==4.0.3
django-allauth==0.37.1
django-cors-headers==2.4.0
django-filter==21.1
django-rest-auth==0.9.3
django-session-timeout==0.1.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.8.2
filelock==3.6.0
Flask==2.0.3
Flask-WTF==1.0.0
greenlet==1.1.2
httplib2==0.11.3
idna==2.7
itsdangerous==2.1.1
Jinja2==3.0.3
Js2Py==0.71
louis==3.5.0
MarkupSafe==2.1.0
mysql-connector==2.2.9
mysqlclient==2.1.0
oauthlib==2.1.0
olefile==0.45.1
onboard==1.4.1
passlib==1.7.4
pexpect==4.2.1
Pillow==5.1.0
platformdirs==2.5.2
protobuf==3.19.4
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.20
PyGObject==3.26.1
pyjsparser==2.7.1
PyJWT==1.7.1
PyMySQL==1.0.2
PyPDF2==1.27.12
python-apt===1.6.5-ubuntu0.7-elementary19-ubuntu18.04.1
python3-openid==3.1.0
pytz==2018.5
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
reportlab==3.4.0
requests==2.20.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
six==1.11.0
SQLAlchemy==1.4.32
sqlparse==0.3.0
systemd-python==234
tzdata==2021.5
tzlocal==4.1
ubuntu-drivers-common==0.0.0
urllib3==1.23
validate-email==1.3
virtualenv==20.14.1
Werkzeug==2.0.3
wrapt==1.11.1
WTForms==3.0.1
xkit==0.0.0
xlwt==1.3.0
